yjw2288/ssiggle/ssiggle/src/main/java/kr/ac/jejuuniv/controller/ModifyUserController.java
package kr.ac.jejuuniv.controller;

import kr.ac.jejuuniv.model.User;
import kr.ac.jejuuniv.service.UserService;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.ui.ModelMap;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;

@Controller
@RequestMapping("/modifyuser")
public class ModifyUserController {
	@Autowired
	private UserService userService;

	@RequestMapping(method = RequestMethod.GET)
	public String action(@RequestParam String id, ModelMap modelMap) {
		User user = userService.findUserById(id);
		user.setPassword("");

		modelMap.addAttribute("user", user);
		return "usermodifypage";
	}
}
